+++ builder_pkgs/mockito/lib/mockito.dart
@@ -28,6 +28,12 @@ export 'src/mock.dart'
         Expectation,
         PostExpectation,
 
+        // -- Mockito 3.0 compatibility
+        anyNamed,
+        captureAnyNamed,
+        typedArgThat,
+        typedCaptureThat,
+
         // -- verification
         verify,
         verifyInOrder,