+++ builder_pkgs/mockito/test/mockito_test.dart
@@ -1080,4 +1080,53 @@ void main() {
       expect(() => mock.methodWithoutArgs(), returnsNormally);
     });
   });
+
+  group("the intermediate API", () {
+    test("should mock method with multiple named args and matchers", () {
+      when(mock.methodWithTwoNamedArgs(typed(any), y: anyNamed('y')))
+          .thenReturn("x y");
+      when(mock.methodWithTwoNamedArgs(typed(any), z: anyNamed('z')))
+          .thenReturn("x z");
+      expect(mock.methodWithTwoNamedArgs(42), isNull);
+      expect(mock.methodWithTwoNamedArgs(42, y: 18), equals("x y"));
+      expect(mock.methodWithTwoNamedArgs(42, z: 17), equals("x z"));
+      expect(mock.methodWithTwoNamedArgs(42, y: 18, z: 17), isNull);
+      when(mock.methodWithTwoNamedArgs(typed(any),
+              y: anyNamed('y'), z: anyNamed('z')))
+          .thenReturn("x y z");
+      expect(mock.methodWithTwoNamedArgs(42, y: 18, z: 17), equals("x y z"));
+    });
+
+    test("should capture multiple named args", () {
+      mock.methodWithTwoNamedArgs(42, y: 18);
+      mock.methodWithTwoNamedArgs(42, z: 17);
+      var verifiedY = verify(
+          mock.methodWithTwoNamedArgs(typed(any), y: captureAnyNamed('y')));
+      expect(verifiedY.captured.single, equals(18));
+      var verifiedZ = verify(
+          mock.methodWithTwoNamedArgs(typed(any), z: captureAnyNamed('z')));
+      expect(verifiedZ.captured.single, equals(17));
+    });
+
+    test("should mock method with named, typed arg matcher and an arg matcher",
+        () {
+      when(mock.typeParameterizedNamedFn(typed(any), [43],
+              y: typed(any, named: "y"),
+              z: typedArgThat(contains(45), named: 'z')))
+          .thenReturn("A lot!");
+      expect(mock.typeParameterizedNamedFn([42], [43], y: [44], z: [45]),
+          equals("A lot!"));
+    });
+
+    test("should capture multiple named args", () {
+      mock.methodWithTwoNamedArgs(42, y: 18);
+      mock.methodWithTwoNamedArgs(42, z: 17);
+      var verifiedY = verify(mock.methodWithTwoNamedArgs(typed(any),
+          y: typedCaptureThat(lessThan(75), named: 'y')));
+      expect(verifiedY.captured.single, equals(18));
+      var verifiedZ = verify(mock.methodWithTwoNamedArgs(typed(any),
+          z: typedCaptureThat(lessThan(75), named: 'z')));
+      expect(verifiedZ.captured.single, equals(17));
+    });
+  });
 }